cleanup(core): move the repo key git operations into git-utils
The git-root-relative path, shallow check, and first-commit lookup are git concerns, so they join the other standalone helpers in git-utils rather than living in repo-key. GitRepository.getGitRootPath now delegates to an exported function instead of holding a second copy of the rev-parse call, and the relative path uses the file's existing sep/posix imports. repo-key keeps only identity selection and hashing.
